Cooking works, change color based on real model

+++ Assets/Scripts/Properties.cs
@@ -7,15 +7,15 @@ public class Properties : MonoBehaviour
     public GameObject unfinished;
     public GameObject finished;
 
-    public int heatTime;
-    private int heated;
+    public float heatTime;
+    private float heated;
     // Start is called before the first frame update
     void Start()
     {
         unfinished = transform.GetChild(0).gameObject;
         finished = transform.GetChild(1).gameObject;
         finished.SetActive(false);
-        heated = 0;
+        heated = 0f;
     }
 
     // Update is called once per frame
@@ -24,12 +24,12 @@ public class Properties : MonoBehaviour
         
     }
 
-    public int getHeat()
+    public float getHeat()
     {
         return heated;
     }
 
-    public void setHeat(int newHeat)
+    public void setHeat(float newHeat)
     {
         heated = newHeat;
     }
